+++ interpreter/Interpreter1/src/Node.java
@@ -9,11 +9,10 @@ public abstract class Node {
 	/**
 	 * The function which all other nodes implement to perform whatever evaluation is necessary for their type
 	 * @param env - the stored variables and such
-	 * @return - the value of the evaluation, an int in our math language
+	 * @return - the value of the evaluation, a double in our math language
 	 * @throws EvalException 
 	 */
-    public int eval(Environment env) throws EvalException {
+    public double eval(Environment env) throws EvalException {
 	throw new EvalException(pos,"cannot eval() node!");
     }
-
 }
