+++ interpreter/Interpreter1/src/NodeFactExpr.java
@@ -20,7 +20,7 @@ public class NodeFactExpr extends NodeFact {
 	 * @return - either the value of the term or the value of whatever may be contained in it: an id, number or expression 
 	 * @throws EvalException 
 	 */
-    public int eval(Environment env) throws EvalException {
+    public double eval(Environment env) throws EvalException {
 	return expr.eval(env);
     }
