+++ interpreter/Interpreter1/src/NodeFactNum.java
@@ -38,7 +38,7 @@ public class NodeFactNum extends NodeFact
 	 * @return
 	 * @throws EvalException 
 	 */
-	public int eval(Environment env) throws EvalException
+	public double eval(Environment env) throws EvalException
 	{
 		return num.eval(env);
 	}
