+++ interpreter/Interpreter1/src/NodeMulop.java
@@ -21,7 +21,7 @@ public class NodeMulop extends Node
 	 * @return - the product or quotient of the numbers
 	 * @throws EvalException 
 	 */
-	public int op(int o1, int o2) throws EvalException
+	public double op(double o1, double o2) throws EvalException
 	{
 		if (mulop.equals("*"))
 		{
