+++ interpreter/Interpreter1/src/NodeStmt.java
@@ -23,7 +23,7 @@ public class NodeStmt extends Node
 	 * @return - perhaps some kind of id from the environment
 	 * @throws EvalException 
 	 */
-	public int eval(Environment env) throws EvalException
+	public double eval(Environment env) throws EvalException
 	{
 		return assn.eval(env);
 	}
